Clicking on any of the links in the nav closes it
Why? Because when I click on a link, the react app routes to that link,
but the side nav stays, it doesn't get closed

I did it by making a `CustomLink` component within the `Nav` component,
and replaced all the links with these passing all the props. This custom
Link, has an onClick method which closes the side nav